[refactor]: Constants 상수들 객체 내로 이동, Menu validator 이동
