--- src/main/java/calculator/global/ErrorResponse.java
@@ -1,8 +1,0 @@
-package calculator.global;
-
-public class ErrorResponse {
-    public static final String MENU_INPUT_ERROR = "메뉴가 올바르게 입력되지 않았습니다.";
-    public static final String OPERATOR_INPUT_ERROR = "연산자 입력에 오류가 발생했습니다.";
-    public static final String INVALID_OPERAND = "올바른 피연산자가 아닙니다.";
-    public static final String INVALID_INPUT_EXPRESSION = "올바른 수식 표현이 아닙니다.";
-}
