--- src/main/java/calculator/global/InputConstants.java
@@ -1,16 +1,0 @@
-package calculator.global;
-
-public class InputConstants {
-    public static final String CHOICE_PROMPT = "\n선택 : ";
-    public static final int FIRST_INDEX = 0;
-    public static final int MENU_INPUT_LENGTH = 1;
-
-    public static final char REQUEST_EXIT = '0';
-    public static final char REQUEST_VIEW_CALCULATION_RESULT = '1';
-    public static final char REQUEST_CALCULATION = '2';
-
-    public static final String OPERATOR_REGEX = "[-*/+]";
-    public static final String OPERAND_REGEX = "^\\d+$";
-
-    public static final int MINIMUM_OPS = 3;
-}
