--- src/main/java/calculator/global/Priority.java
@@ -1,6 +1,0 @@
-package calculator.global;
-
-public class Priority {
-    public static final Integer LOW = 0;
-    public static final Integer HIGH = 1;
-}
